Fixed compile error with Pair class

+++ WEB-INF/classes/util/Pair.java
@@ -33,7 +33,7 @@ public class Pair implements Comparable<Double> {
 		if (this.getKey() > that) return AFTER;
 		if (this.getKey() == that) return EQUAL;
 
-		return Integer.MIN_INTEGER;
+		return (int) Integer.MIN_INTEGER;
 	}
 	
 }
